[css-flexbox] min-height:auto not updated after an image loads when the image has a specified height and width. https://bugs.webkit.org/show_bug.cgi?id=210475
Reviewed by Javier Fernandez.

Source/WebCore:

The flex layout algorithm needs the item's intrinsic size even if it has a specified size, that's
why the flex item size should be recomputed in the event of changes in its intrinsic size.

* rendering/RenderReplaced.cpp:
(WebCore::RenderReplaced::setNeedsLayoutIfNeededAfterIntrinsicSizeChange): set the needs layout flag
also for replaced elements which happen to be flex items.

LayoutTests:

* TestExpectations: Removed two tests that are passing now.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@265858 268f45cc-cd09-0410-ab3c-d52691b4dbfc