A PiP window is closed when the video element is removed from DOM https://bugs.webkit.org/show_bug.cgi?id=215594
Reviewed by Eric Carlson.

Source/WebCore:

Test: media/remove-video-element-in-pip-from-document.html

* html/HTMLMediaElement.cpp:
(WebCore::HTMLMediaElement::pauseAfterDetachedTask):
When a video is playing in the picture-in-picture mode, do not pause it or exit the
picture-in-picture mode when detaching the video element from the DOM.
(WebCore::HTMLMediaElement::mediaPlayerRenderingCanBeAccelerated):
Always return true when the video is playing in the picture-in-picture mode.
* rendering/RenderVideo.cpp:
(WebCore::RenderVideo::willBeDestroyed):
Don't hide the player when destroying the renderer of an inline video.

LayoutTests:

* TestExpectations:
* media/remove-video-element-in-pip-from-document-expected.txt: Added.
* media/remove-video-element-in-pip-from-document.html: Added.
* platform/ipad/TestExpectations:
* platform/mac-wk2/TestExpectations:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@265904 268f45cc-cd09-0410-ab3c-d52691b4dbfc