Trying to lookup when WebView is in a popover causes process to hang. Fix for Legacy WebView. https://bugs.webkit.org/show_bug.cgi?id=215792
Reviewed by Tim Horton.

Source/WebKit:

Update code based on Darin's suggestion for mirror fix in legacy webkit.

* UIProcess/mac/WebContextMenuProxyMac.mm:
(WebKit::WebContextMenuProxyMac::getContextMenuFromItems):

Source/WebKitLegacy/mac:

Fix for https://bugs.webkit.org/show_bug.cgi?id=214773 also needed in legacy WebView.
The Lookup framework does not populate the menus that it vends with the option to 'lookup'
words that are selected in popovers. WebKit should follow the pattern and not put the
item in the menu as Lookup is not able to handle this situation. Also if Lookup is disabled
via defaults, we should not show it in the menu.

* WebView/WebHTMLView.mm:
(customMenuFromDefaultItems):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@266143 268f45cc-cd09-0410-ab3c-d52691b4dbfc