Flickering on sedona.dev https://bugs.webkit.org/show_bug.cgi?id=215141
Reviewed by Darin Adler.

Source/WebCore:

Test: fast/canvas/webgl/compositing-without-drawing.html

Our logic to determine if a canvas needs to be "repainted"
was over-zealous for WebGL. We were marking any context
that called draw commands as dirty, but they could in fact
be rendering to an offscreen texture/framebuffer. Then, when
it came time to composite, we'd happily swap buffers and
show something that had never been rendered to.

The fix is simply to ignore any of the dirtying notifications
when we are not bound to the default (canvas) framebuffer.

* html/canvas/WebGLRenderingContextBase.cpp:
(WebCore::WebGLRenderingContextBase::markContextChangedAndNotifyCanvasObserver): Only
mark if we're rendering to the default framebuffer.

LayoutTests:

Test that serves an animation frame that touches
WebGL, but not in a way that requires a recomposite.

* fast/canvas/webgl/compositing-without-drawing-expected.html: Added.
* fast/canvas/webgl/compositing-without-drawing.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@266189 268f45cc-cd09-0410-ab3c-d52691b4dbfc