REGRESSION(r266187): ARGUMENT BAD: WebIconUtilities.mm(138) : RetainPtr<UIImage> WebKit::iconForFile(NSURL *) file, [file isFileURL] https://bugs.webkit.org/show_bug.cgi?id=215931
Patch by Alex Christensen <achristensen@webkit.org> on 2020-08-28
Reviewed by Wenson Hsieh.

It's possible for JavaScript to make a filename that doesn't convert to a file URL, such as the empty string.
When this happens, we would've returned null for the icon later anyways, but adding a null check prevents an assertion.
Covered by 7 existing tests which would have hit that assertion.

* WebProcess/WebCoreSupport/ios/WebChromeClientIOS.mm:
(WebKit::WebChromeClient::createIconForFiles):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@266287 268f45cc-cd09-0410-ab3c-d52691b4dbfc