[JSC] for-of / for-in left-hand-side target should be simple-assignment-target https://bugs.webkit.org/show_bug.cgi?id=215969
Reviewed by Ross Kirsling.

JSTests:

* stress/for-of-in-non-reference.js: Added.
(testSyntax):
(testSyntaxError):
(testSyntaxError.t):
(testSyntaxError.async t):
(testSyntax.async t):
* test262/expectations.yaml:

Source/JavaScriptCore:

Left-hand-side of `for-in`, `for-of`, and `for-await-of` should be simple assignment target[1]
if the target is not declaration and not destructuring pattern.

[1]: https://tc39.es/ecma262/#sec-for-in-and-for-of-statements-static-semantics-early-errors

* parser/Parser.cpp:
(JSC::Parser<LexerType>::parseForStatement):
* parser/SyntaxChecker.h:
(JSC::SyntaxChecker::createCommaExpr): Should return CommaExpr to align it to ASTBuilder.
(JSC::SyntaxChecker::appendToCommaExpr):
(JSC::SyntaxChecker::appendStatement):
(JSC::SyntaxChecker::combineCommaNodes): Deleted since it is not used.

LayoutTests:

Fix for-in / for-of parser behavior.

* js/basic-strict-mode-expected.txt:
* js/parser-syntax-check-expected.txt:
* js/script-tests/parser-syntax-check.js:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@266324 268f45cc-cd09-0410-ab3c-d52691b4dbfc