AudioParam.linearRampToValueAtTime() / exponentialRampToValueAtTime() have no effect when there is no preceding event https://bugs.webkit.org/show_bug.cgi?id=216284
Reviewed by Darin Adler.

LayoutTests/imported/w3c:

* web-platform-tests/webaudio/the-audio-api/the-audioparam-interface/k-rate-connections-expected.txt:
Rebaseline test that is now failing because of the fix. The test was not actually running as intended because
it relies on AudioParam.linearRampToValueAtTime(), which was not working until now. I have investigated this
failure and it is actually due to a bug in our AudioBus::copyWithGainFrom() implementation. I will land a
follow-up fix for this.

* web-platform-tests/webaudio/the-audio-api/the-audioparam-interface/k-rate-constant-source-expected.txt:
Rebaseline test now that output is different (still passing).

* web-platform-tests/webaudio/the-audio-api/the-oscillatornode-interface/detune-limiting-expected.txt:
Rebaseline test that is passing thanks to the fix.

Source/WebCore:

AudioParam.linearRampToValueAtTime() / exponentialRampToValueAtTime() have no effect when there is no preceding event
in the timeline. This is incorrect. We should insert an implicit SetValue event in the timeline if the ramp event is
the first one.

No new tests, rebaselined existing test.

* Modules/webaudio/AudioParam.cpp:
* Modules/webaudio/AudioParamTimeline.cpp:
(WebCore::AudioParamTimeline::linearRampToValueAtTime):
(WebCore::AudioParamTimeline::exponentialRampToValueAtTime):
* Modules/webaudio/AudioParamTimeline.h:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@266788 268f45cc-cd09-0410-ab3c-d52691b4dbfc