[CG] REGRESSION (Big Sur): A GIF image with a finite loopCount loops an extra cycle https://bugs.webkit.org/show_bug.cgi?id=216018 <rdar://problem/68304035>
Patch by Said Abou-Hallawa <sabouhallawa@apple.com> on 2020-09-10
Reviewed by Tim Horton.

Source/WebCore:

Remove the extra 'one' we used to add to the GIF loopCount since it is
now added by the underlying frameworks. But make sure we are compatible
with the older versions of macOS and iOS.

* platform/graphics/cg/ImageDecoderCG.cpp:
(WebCore::ImageDecoderCG::repetitionCount const):

Source/WTF:

Add a new macro for the new accurate behavior of CGImageSource.

Unrelated change: Fix the conditions for enabling the WebP images.

* wtf/PlatformHave.h:

LayoutTests:

* platform/mac/TestExpectations:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@266827 268f45cc-cd09-0410-ab3c-d52691b4dbfc